Enable layered encryption for tunnel building (#70)
Apart from enabling onion encryption for tunnel extend/connect message
this commit rethink the idea of wrapping messages that should be encrypted.
It introduces a "Relay" message that is capable of transporting encrypted
payloads of any type derived from TunnelRelayPayload or raw bytes. This
unifies the encryption for all classes (no Tunnel*EncryptedMessage anymore)
and simplifies the code.